move: Main 위치 변경하기
